requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0          # async postgres driver
alembic==1.13.1          # migrations
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0   # JWT
bcrypt==4.1.3            # PIN hashing
python-multipart==0.0.9  # upload slip image